Move toast to it's namespace
